Refactoring models and serializers into named versions
Currently we have two different sets of API endpoints, the second-one had
a naming scheme with _new for files, and suffix New for classes.

On dataset level, these different sets are split into versions v1 and v2
