+++ web/panorama/datasets/panoramas/tests/factories.py
@@ -1,11 +1,12 @@
 import factory
 
+import datasets.panoramas.v1.models
 from .. import models
 
 
 class PanoramaFactory(factory.DjangoModelFactory):
     class Meta:
-        model = models.Panorama
+        model = datasets.panoramas.v1.models.Panorama
 
 
 class TrajectFactory(factory.DjangoModelFactory):
