+++ web/panorama/datasets/panoramas/v2/hal_serializer.py
