+++ web/panorama/datasets/panoramas/v2/models.py
