+++ web/panorama/panorama/tasks/tests/test_render.py
@@ -9,7 +9,7 @@ import factory.fuzzy
 from django.contrib.gis.geos import Point
 from django.utils.timezone import utc as UTC_TZ
 
-from datasets.panoramas.models import Panorama
+from datasets.panoramas.v1.models import Panorama
 from datasets.panoramas.tests import factories
 from panorama.transform.tests.test_img_file import mock_get_raw_pano
 from .. workers.render_pano import PanoRenderer
